# encoding.base64: speed up encoding and decoding