--- a/vlib/encoding/base64/base64.v
+++ b/vlib/encoding/base64/base64.v
@@ -223,37 +223,29 @@ fn decode_from_buffer(dest &byte, src &byte, src_len int) int {
 		mut si := 0
 
 		for src_len - si >= 8 {
-			dn := assemble64(
-				byte(base64.index[d[si+0]]),
-				byte(base64.index[d[si+1]]),
-				byte(base64.index[d[si+2]]),
-				byte(base64.index[d[si+3]]),
-				byte(base64.index[d[si+4]]),
-				byte(base64.index[d[si+5]]),
-				byte(base64.index[d[si+6]]),
-				byte(base64.index[d[si+7]]))
+			dn := assemble64(byte(base64.index[d[si + 0]]), byte(base64.index[d[si + 1]]),
+				byte(base64.index[d[si + 2]]), byte(base64.index[d[si + 3]]), byte(base64.index[d[
+				si + 4]]), byte(base64.index[d[si + 5]]), byte(base64.index[d[si + 6]]),
+				byte(base64.index[d[si + 7]]))
 
 			// binary.big_endian_put_u64(mut b[n_decoded_bytes..(n_decoded_bytes+8)], dn)
-//			_ = b[n_decoded_bytes + 7] // bounds check
+			//			_ = b[n_decoded_bytes + 7] // bounds check
 			b[n_decoded_bytes + 0] = byte(dn >> u64(56))
 			b[n_decoded_bytes + 1] = byte(dn >> u64(48))
 			b[n_decoded_bytes + 2] = byte(dn >> u64(40))
 			b[n_decoded_bytes + 3] = byte(dn >> u64(32))
 			b[n_decoded_bytes + 4] = byte(dn >> u64(24))
 			b[n_decoded_bytes + 5] = byte(dn >> u64(16))
-//			b[n_decoded_bytes + 6] = byte(dn >> u64(8))
-//			b[n_decoded_bytes + 7] = byte(dn)
+			//			b[n_decoded_bytes + 6] = byte(dn >> u64(8))
+			//			b[n_decoded_bytes + 7] = byte(dn)
 
 			n_decoded_bytes += 6
 			si += 8
 		}
 
 		for src_len - si >= 4 {
-			dn := assemble32(
-				byte(base64.index[d[si+0]]),
-				byte(base64.index[d[si+1]]),
-				byte(base64.index[d[si+2]]),
-				byte(base64.index[d[si+3]]))
+			dn := assemble32(byte(base64.index[d[si + 0]]), byte(base64.index[d[si + 1]]),
+				byte(base64.index[d[si + 2]]), byte(base64.index[d[si + 3]]))
 
 			// binary.big_endian_put_u32(mut b []byte, v u32)
 			_ = b[n_decoded_bytes + 3] // bounds check