## Replacing the previous implementation

This PR started as the result of me writing the V core-util `base64` implementation and noticing that the library code wasn't exactly fast.

With this PR I ripped out the previous implementation and replaced it with GO's implementation. Both projects are MIT licensed and I mentioned that I copied and adapted the code from the GO source in the beginning of the file - as it is also done in the crypto modules.

## Benchmarks and Comparison

Testing was performed on these files, which are the encoded / decoded form of each other, respectively:
```
523M -rw-rw-r-- 1 basti basti 523M Aug  4 14:16 /tmp/base64data.txt
388M -rw-rw-r-- 1 basti basti 388M Aug  4 14:07 /tmp/randomtext.txt
```

Important excerpts from the profiling logs: (Here I used the V implementation of the base64 core-util)

#### Encoding:

- Old Implementation:
```
26422       1366.857ms          51732ns encoding__base64__encode_in_buffer 
26422       1365.009ms          51662ns encoding__base64__encode_from_buffer 
```
- New Implementation:

```
26422        925.792ms          35039ns encoding__base64__encode_in_buffer 
26422        923.648ms          34958ns encoding__base64__encode_from_buffer 
```
#### Decoding:

- Old Implementation:
```
33029      45320.907ms        1372155ns os__File_read_bytes_into_newline
33028      42052.360ms        1273234ns encoding__base64__decode_in_buffer
```
- New Implementation:
```
33029      43124.751ms        1305663ns os__File_read_bytes_into_newline
33028      45464.197ms        1376535ns encoding__base64__decode_in_buffer 
33028      45461.255ms        1376446ns encoding__base64__decode_from_buffer 
```

As you have probably noticed, the decoding function seems to be very slow. I don't think the benchmark is really helpful in this situation.
I think the problem here is that the file reading function is much slower, if not run with the `-prod` flag. Sadly benchmarking does not support running optimized code to my knowledge.
To sum it up, I don't believe the benchmark is showing accurate/helpful results for the decoding functions.

So to provide a more accurate example, I compiled the code with the `-prod` flag and just used `time` to get results.

```
// new base64:
time ./base64 -d /tmp/base64data.txt > /dev/null 
real 2.20s, user 2.06s, sys 0.14s, mem 2132KB, cpu 99%

// old base64:
time ./base64 -d /tmp/base64data.txt > /dev/null 
real 2.64s, user 2.50s, sys 0.14s, mem 2192KB, cpu 100%
```

So there is definitely a speed up by using the newer code.

#### Remaining problems

~~I had to copy code from the binary module to make it work on the buffers, as the buffers are not slices/array, but basically just C style pointers. I could have used slices, but all the other functions did not do that either, so for now I stuck with function parameters in the way it was implemented before.~~

~~The `base64_memory_test.v` fails and I cannot pinpoint the problem. Seems to be related to the use of malloc for the buffer. So it seems like it's a problem with the C code of malloc, but it works on master. Please help.~~