@spytheman 2021-08-04
> The `base64_memory_test.v` fails and I cannot pinpoint the problem. Seems to be related to the use of malloc for the buffer. So it seems like it's a problem with the C code of malloc, but it works on master. Please help.
> 
> ```
> base64_memory_test: malloc.c:2379: sysmalloc: Assertion `(old_top == initial_top (av) && old_size == 0) || ((unsigned long) (old_size) >= MINSIZE && prev_inuse (old_top) && ((unsigned long) old_end & (pagesize - 1)) == 0)' failed.
> 7f2f4cae818b : at ???: RUNTIME ERROR: abort() called
> ```

Compiling with `./v -g -keepc -cc gcc vlib/encoding/base64/base64_memory_test.v`, 
then running under valgrind with: `valgrind -s vlib/encoding/base64/base64_memory_test` produces:
```
....
==11176== 
==11176== 
==11176== 1000 errors in context 6 of 6:
==11176== Invalid read of size 1
==11176==    at 0x4437A9: encoding__base64__decode_from_buffer (base64.v:231)
==11176==    by 0x443264: encoding__base64__decode_in_buffer (base64.v:129)
==11176==    by 0x44B29C: main__test_long_encoding (base64_memory_test.v:25)
==11176==    by 0x4533F5: main (base64_memory_test.tmp.c:1000363)
==11176==  Address 0x4b2f329 is 1 bytes after a block of size 3,000 alloc'd
==11176==    at 0x483B7F3: malloc (vg_replace_malloc.c:309)
==11176==    by 0x4203FD: _v_malloc (builtin.c.v:285)
==11176==    by 0x44B26D: main__test_long_encoding (base64_memory_test.v:23)
==11176==    by 0x4533F5: main (base64_memory_test.tmp.c:1000363)
==11176== 
==11176== ERROR SUMMARY: 3003 errors from 6 contexts (suppressed: 0 from 0)
```


```