@spytheman 2021-08-04
this fixes it for me 🤔 
```v
//          _ = b[n_decoded_bytes + 7] // bounds check
            b[n_decoded_bytes + 0] = byte(dn >> u64(56))
            b[n_decoded_bytes + 1] = byte(dn >> u64(48))
            b[n_decoded_bytes + 2] = byte(dn >> u64(40))
            b[n_decoded_bytes + 3] = byte(dn >> u64(32))
            b[n_decoded_bytes + 4] = byte(dn >> u64(24))
            b[n_decoded_bytes + 5] = byte(dn >> u64(16))
//          b[n_decoded_bytes + 6] = byte(dn >> u64(8))
//          b[n_decoded_bytes + 7] = byte(dn)
```