@spytheman 2021-08-04
if we already have the decoded part in an u64 `dn`, could not we just write it directly, instead of writing it byte by byte?