@schicho 2021-08-05
> If it needs to be separate bytes, just create a union. Write `dn` directly, then read it out via the byte array. Even better... use the `u64` part of the union _as_ `dn`, and you don't have to do any extra conversions anywhere.

I like this solution and will now look into it.