@schicho 2021-08-05
I've addressed the reviews and have now decided to implement the decoding as @JalonSolov suggested.