@spytheman 2021-08-06
(rebased over current master)