@JalonSolov 2021-08-06
Are there new timings of the old vs new?