@schicho 2021-08-07
I currently do not have access to a PC over the weekend, so I can't benchmark right now.

I can look into it, if still required, when I have access to a PC again.